requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
PyMuPDF>=1.23.0
pytesseract>=0.3.10
pillow>=10.0.0
numpy>=1.26.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
jupyter>=1.0.0